+++ .coveragerc
@@ -18,7 +18,7 @@ ignore_errors = True
 omit =
     ocrd_models/ocrd_models/ocrd_page_generateds.py
     */setup.py
-    ocrd/ocrd/cli/*
-    ocrd/ocrd/decorators.py
+    ocrd/ocrd/cli*
+    # ocrd/ocrd/decorators.py
     ocrd/ocrd/task_sequence.py
     tests/*
